chore(data-service): setup typescript for data-service COMPASS-5006
