+++ packages/compass-components/.prettierignore
@@ -1,1 +1,2 @@
 lib
+.nyc_output
