+++ packages/data-service/.depcheckrc
@@ -1,1 +1,5 @@
-ignores: []
+ignores: [
+  "ts-node",
+  "@mongodb-js/prettier-config-compass",
+  "@mongodb-js/tsconfig-compass"
+]
