--- packages/data-service/.eslintrc
@@ -1,25 +1,0 @@
-{
-  "env": {
-    "mocha": true,
-    "node": true,
-    "es6": true
-  },
-  "rules": {
-    "strict": "off",
-    "no-undef": "error",
-    "valid-jsdoc": "off",
-    "no-prototype-builtins": "off",
-    "mocha/no-exclusive-tests": "error",
-    "no-unused-vars": "error"
-  },
-  "plugins": [
-    "mocha"
-  ],
-  "extends": [
-    "eslint:recommended",
-    "mongodb-js/node"
-  ],
-  "parserOptions": {
-    "ecmaVersion": 11
-  }
-}
