+++ packages/data-service/.prettierignore
@@ -1,0 +1,2 @@
+lib
+.nyc_output
