+++ packages/data-service/package.json
@@ -19,15 +19,34 @@
   "keywords": [
     "mongodb-js"
   ],
+  "main": "lib/index.js",
+  "exports": {
+    "require": "./lib/index.js",
+    "import": "./lib/.esm-wrapper.mjs"
+  },
+  "types": "./lib/index.d.ts",
+  "files": [
+    "lib",
+    "package.json",
+    "README.md"
+  ],
   "scripts": {
-    "pretest": "mongodb-runner start --port=27018",
-    "test": "mocha",
-    "posttest": "mongodb-runner stop --port=27018",
     "check": "npm run lint && npm run depcheck",
+    "clean": "rimraf lib",
+    "precompile": "npm run clean",
+    "compile": "npm run compile-ts && gen-esm-wrapper . ./lib/.esm-wrapper.mjs",
+    "compile-ts": "tsc -p tsconfig.json",
+    "cover": "nyc npm run test",
     "depcheck": "depcheck",
-    "lint": "eslint \"./{lib,test}/**/*.js\" \"./*.js\"",
+    "lint": "eslint . && prettier --check .",
+    "prepare": "npm run compile",
+    "pretest": "npm run compile && mongodb-runner start --port=27018",
+    "test": "npm run lint && nyc mocha --colors --timeout 15000 -r ts-node/register {src,test}/**/*.spec.*",
     "test-ci": "npm run test",
-    "posttest-ci": "node ../../scripts/killall-mongo.js"
+    "test-check-ci": "npm run check && npm test",
+    "posttest": "mongodb-runner stop --port=27018",
+    "posttest-ci": "node ../../scripts/killall-mongo.js",
+    "reformat": "prettier --write ."
   },
   "pre-commit": [
     "check"
@@ -48,18 +67,24 @@
     "mongodb-url": "^3.0.3"
   },
   "devDependencies": {
+    "@mongodb-js/eslint-config-compass": "^0.0.1",
+    "@mongodb-js/prettier-config-compass": "^0.0.1",
+    "@mongodb-js/tsconfig-compass": "^0.0.1",
     "bson": "^4.4.1",
     "chai": "^4.2.0",
     "depcheck": "^1.4.1",
     "eslint": "^7.25.0",
-    "eslint-config-mongodb-js": "^5.0.3",
-    "eslint-plugin-mocha": "^8.0.0",
-    "mocha": "^8.2.1",
+    "mocha": "^8.4.0",
     "mock-require": "^3.0.3",
     "mongodb": "^4.1.0",
     "mongodb-connection-model": "^21.6.0",
     "mongodb-runner": "^4.8.1",
+    "nyc": "^15.0.0",
+    "prettier": "2.3.2",
+    "rimraf": "^3.0.2",
     "sinon": "^9.2.3",
-    "sinon-chai": "^3.5.0"
+    "sinon-chai": "^3.5.0",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   }
 }
