+++ packages/data-service/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@mongodb-js/tsconfig-compass/tsconfig.common.json",
+  "compilerOptions": {
+    "allowJs": true,
+    "outDir": "lib"
+  },
+  "include": ["src/**/*"],
+  "exclude": ["./src/**/*.spec.*"]
+}
