+++ packages/prettier-config-compass/README.md
@@ -1,0 +1,5 @@
+# @mongodb-js/prettier-config-compass
+
+> Shared [prettier][prettier] configuration for Compass packages.
+
+[prettier]: https://prettier.io/
