+++ packages/prettier-config-compass/index.json
@@ -1,0 +1,3 @@
+{
+  "singleQuote": true
+}
