+++ packages/tsconfig-compass/.prettierrc.json
@@ -1,1 +1,1 @@
-{}
+"@mongodb-js/prettier-config-compass"
